requirements.txt
flask
flask-cors
langchain
langchain-community
langchain-core
chromadb
sentence-transformers
google-generativeai
python-dotenv
pypdf
tiktoken
markdown
langchain-google-genai